backend/pyproject.toml
[tool.poetry]
name = "api"
version = "0.0.1"
description = "FASTAPI server that hosts api and manages queue/worker job start."
authors = ["Team Jataware <joel@jataware.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
pydantic-settings = "^2.1.0"
openai = "^1.14.2"
rq = "1.10.0"
uvicorn = "^0.24.0.post1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
elasticsearch = "^8.13.0"
markdown = "^3.6"
pillow = "^10.3.0"

[tool.poetry.group.workers.dependencies]
selenium = "^4.18.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"